+++ x/metoken/keeper/intest/msg_server_test.go
@@ -1,9 +1,7 @@
-//go:build experimental
-// +build experimental
-
 package intest
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/golang/mock/gomock"
@@ -250,6 +248,9 @@ func TestMsgServer_Swap_NonStableAssets_DiffExponents(t *testing.T) {
 			require.NoError(err)
 
 			// verify the outputs of swap function
+			if tc.name == "valid - swap 1.435125562353141231 ETH" {
+				fmt.Printf("\n")
+			}
 			resp, err := msgServer.Swap(ctx, msg)
 			require.NoError(err, tc.name)
 
@@ -612,6 +613,7 @@ func TestMsgServer_Swap_Depegging(t *testing.T) {
 		app.BankKeeper,
 		app.LeverageKeeper,
 		oracleMock,
+		app.UGovKeeperB.EmergencyGroup,
 	)
 	app.MetokenKeeperB = kb
 	msgServer = keeper.NewMsgServerImpl(app.MetokenKeeperB)
@@ -708,6 +710,7 @@ func TestMsgServer_Swap_Depegging(t *testing.T) {
 		app.BankKeeper,
 		app.LeverageKeeper,
 		oracleMock,
+		app.UGovKeeperB.EmergencyGroup,
 	)
 	app.MetokenKeeperB = kb
 	msgServer = keeper.NewMsgServerImpl(app.MetokenKeeperB)
@@ -898,11 +901,12 @@ func verifySwap(
 
 	assetExponentFactorVsMeToken, err := metoken.ExponentFactor(assetPrice.Exponent, prices.Exponent)
 	assert.NilError(t, err)
+	rate := exchangeRate.Mul(assetExponentFactorVsMeToken)
 
 	// expected_metokens = amount_to_swap * exchange_rate * exponent_factor
 	expectedMeTokens := sdk.NewCoin(
 		meTokenDenom,
-		exchangeRate.MulInt(amountToSwap).Mul(assetExponentFactorVsMeToken).TruncateInt(),
+		rate.MulInt(amountToSwap).TruncateInt(),
 	)
 
 	// calculating reserved and leveraged
@@ -1350,6 +1354,7 @@ func TestMsgServer_Redeem_Depegging(t *testing.T) {
 		app.BankKeeper,
 		app.LeverageKeeper,
 		oracleMock,
+		app.UGovKeeperB.EmergencyGroup,
 	)
 	app.MetokenKeeperB = kb
 	msgServer = keeper.NewMsgServerImpl(app.MetokenKeeperB)
@@ -1446,6 +1451,7 @@ func TestMsgServer_Redeem_Depegging(t *testing.T) {
 		app.BankKeeper,
 		app.LeverageKeeper,
 		oracleMock,
+		app.UGovKeeperB.EmergencyGroup,
 	)
 	app.MetokenKeeperB = kb
 	msgServer = keeper.NewMsgServerImpl(app.MetokenKeeperB)
@@ -1861,8 +1867,12 @@ func TestMsgServer_GovUpdateRegistry(t *testing.T) {
 	}{
 		{
 			"invalid authority",
-			metoken.NewMsgGovUpdateRegistry("invalid_authority", nil, nil),
-			"invalid_authority",
+			metoken.NewMsgGovUpdateRegistry(
+				"umee156hsyuvssklaekm57qy0pcehlfhzpclclaadwq",
+				nil,
+				[]metoken.Index{existingIndex},
+			),
+			"unauthorized",
 		},
 		{
 			"invalid - empty add and update indexes",